requirements.txt
streamlit>=1.27.0
langchain>=0.0.267
langchain-community>=0.0.10
langchain-huggingface>=0.0.3
langchain-core>=0.1.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
ctransformers>=0.2.27
pypdf>=3.15.1
docx2txt>=0.8
python-dotenv>=1.0.0
unstructured>=0.10.30
unstructured-inference>=0.7.21